Move ropt handling to /utils
Created a separate file rather than flooding _utils.py
